Adopted the query code from github.com/tree-sitter/go-tree-sitter which
is more complete/robust. Refactored it to make it easier to understand
and manage and also to open up the door for end users to be able to
register their own predicate "handlers".

As a result `FilterPredicates()` was removed, the filtering is now
implicit. The number of supported predicates also significantly
increased.

There are BREAKING CHANGES in here, see the updated README.